+++ include/utils/stats.hpp
@@ -11,23 +11,23 @@
 using namespace std;
 
 // Function to return maximum value of vector
-long double vector_max(vector<long double> v);
+double vector_max(vector<double> v);
 
 // Function to return minimum value of vector
-long double vector_min(vector<long double> v);
+double vector_min(vector<double> v);
 
 // Compute mean of a vector
-long double mean(vector<long double> x);
+double mean(vector<double> x);
 
 // Compute mean of a deque
-long double stdev(deque<long double> x);
+double stdev(deque<double> x);
 
 // Compute standard deviation of a list
-long double stdev(deque<long double> x);
+double stdev(deque<double> x);
 
 // Compute the x value for the optimium of a linear regression
-vector<long double> quad_max(vector<long double> x, vector<long double> y);
+vector<double> quad_max(vector<double> x, vector<double> y);
 
-vector<long double> gauss(vector< vector<long double> > A);
+vector<double> gauss(vector< vector<double> > A);
 
 #endif
