+++ src/utils/objective.cpp
@@ -3,15 +3,15 @@
 #define PI 3.141592653589793
 
 //// Define the Griewank function
-long double griewank(vector<long double> x){
+double griewank(vector<double> x){
     // Initialize a couple variable for sums
-    long double fx1 = 0;
-    long double fx2 = 1;
+    double fx1 = 0;
+    double fx2 = 1;
 
     // Step through every dimensions and do a thing
     for(int i=0; i < x.size(); i++){
         fx1 += 360*x[i]*x[i];
-        fx2 *= cos(60*x[i]/sqrt(static_cast <long double> (i+1)));
+        fx2 *= cos(60*x[i]/sqrt(static_cast <double> (i+1)));
     }
 
     // Return a sum of sums
@@ -19,10 +19,10 @@ long double griewank(vector<long double> x){
 }
 
 //// Define the Ackley function
-long double ackley(vector<long double> x){
+double ackley(vector<double> x){
     // Initialize a couple variable for sums
-    long double fx1 = 0;
-    long double fx2 = 0;
+    double fx1 = 0;
+    double fx2 = 0;
     int D = static_cast <int> (x.size());
 
     // Step through every dimensions and do a thing
@@ -36,9 +36,9 @@ long double ackley(vector<long double> x){
 }
 
 //// Define the Rastigrin function
-long double rastigrin(vector<long double> x){
+double rastigrin(vector<double> x){
     // Initialize a couple variable for sums
-    long double fx = 0;
+    double fx = 0;
 
     // Step through every dimensions and do a thing
     // The funky coefficients scale the function so it falls on a [-10 10] range, instead of [-5.12, 5.12] as usual.
